206: move files and refactor out utils
