+++ test/navmesh/triangulation.spec.js
@@ -1,7 +1,6 @@
 import test from 'tape';
 import {
   delaunayTriangulation,
-  calculateNavMesh,
   getDTGraph,
   __RewireAPI__ as TriangulationRewireAPI,
 } from '../../src/navmesh/triangulation';
